#1 [implement] class name change
