+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/MainActivity.kt
@@ -6,7 +6,7 @@ package jp.co.yumemi.android.code_check
 import androidx.appcompat.app.AppCompatActivity
 import java.util.*
 
-class TopActivity : AppCompatActivity(R.layout.activity_top) {
+class TopActivity : AppCompatActivity(R.layout.activity_main) {
 
     companion object {
         lateinit var lastSearchDate: Date
