+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/SearchViewModel.kt
@@ -1,0 +1,82 @@
+/*
+ * Copyright © 2021 YUMEMI Inc. All rights reserved.
+ */
+package jp.co.yumemi.android.code_check
+
+import android.content.Context
+import android.os.Parcelable
+import androidx.lifecycle.ViewModel
+import jp.co.yumemi.android.code_check.TopActivity.Companion.lastSearchDate
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.async
+import kotlinx.coroutines.runBlocking
+import kotlinx.parcelize.Parcelize
+import org.json.JSONObject
+import java.util.*
+
+/**
+ * TwoFragment で使う
+ */
+class SearchViewModel(
+    val context: Context
+) : ViewModel() {
+
+    // 検索結果 - ライブラリの削除に伴いコメントアウト
+//   fun searchResults(inputText: String): List<item> = runBlocking {
+//        val client = HttpClient(Android)
+//
+//        return@runBlocking GlobalScope.async {
+//            val response: HttpResponse = client?.get("https://api.github.com/search/repositories") {
+//                header("Accept", "application/vnd.github.v3+json")
+//                parameter("q", inputText)
+//            }
+//
+//            val jsonBody = JSONObject(response.receive<String>())
+//
+//            val jsonItems = jsonBody.optJSONArray("items")!!
+//
+//            val items = mutableListOf<item>()
+//
+//            /**
+//             * アイテムの個数分ループする
+//             */
+//            for (i in 0 until jsonItems.length()) {
+//                val jsonItem = jsonItems.optJSONObject(i)!!
+//                val name = jsonItem.optString("full_name")
+//                val ownerIconUrl = jsonItem.optJSONObject("owner")!!.optString("avatar_url")
+//                val language = jsonItem.optString("language")
+//                val stargazersCount = jsonItem.optLong("stargazers_count")
+//                val watchersCount = jsonItem.optLong("watchers_count")
+//                val forksCount = jsonItem.optLong("forks_conut")
+//                val openIssuesCount = jsonItem.optLong("open_issues_count")
+//
+//                items.add(
+//                    item(
+//                        name = name,
+//                        ownerIconUrl = ownerIconUrl,
+//                        language = context.getString(R.string.written_language, language),
+//                        stargazersCount = stargazersCount,
+//                        watchersCount = watchersCount,
+//                        forksCount = forksCount,
+//                        openIssuesCount = openIssuesCount
+//                    )
+//                )
+//            }
+//
+//            lastSearchDate = Date()
+//
+//            return@async items.toList()
+//        }.await()
+//    }
+}
+
+@Parcelize
+data class DetailItem(
+    val name: String,
+    val ownerIconUrl: String,
+    val language: String,
+    val stargazersCount: Long,
+    val watchersCount: Long,
+    val forksCount: Long,
+    val openIssuesCount: Long,
+) : Parcelable
